+++ test/runtests.jl
@@ -1,3 +1,4 @@
 include("basetests.jl")
 include("imagetests.jl")
+include("widget-tests.jl")
 include("aquatests.jl")